backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
redis==5.0.3
python-dotenv==1.0.1
gunicorn==21.2.0
